requirements.txt
python-docx>=0.8.11
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
latex2mathml>=3.0.0
